renamed: extrinsic/definitions -> extrinsic/xt_primitives
